Move parseExtension
